refactor(cli): move dev to interface
